Better management of I/O
Do not use I/O in the course of execution, unless a debug flag is on.
The signature of `run_game` cannot be `... -> unit`: that's defeating
the purpose of defining an API, we cannot be fishing for some string
on stdout to programmatically know whether a game is won or not.